# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Portfolio API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /research:
    get:
      operationId: listResearchPapers
      tags: [research]
      summary: List research papers
      responses:
        "200":
          description: Research papers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResearchPaper"
    post:
      operationId: createResearchPaper
      tags: [research]
      summary: Create research paper (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResearchPaperInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchPaper"
        "401":
          description: Unauthorized
  /research/{id}:
    patch:
      operationId: updateResearchPaper
      tags: [research]
      summary: Update research paper (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResearchPaperUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchPaper"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteResearchPaper
      tags: [research]
      summary: Delete research paper (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found
...
components:
  schemas:
...
    ResearchPaper:
      type: object
      required: [id, title, authors, abstract, venue, year, tags]
      properties:
        id:
          type: integer
        title:
          type: string
        authors:
          type: string
        abstract:
          type: string
        venue:
          type: string
        year:
          type: integer
        paperLink:
          type: ["string", "null"]
        tags:
          type: string
    ResearchPaperInput:
      type: object
      required: [title, authors, abstract, venue, year, tags]
      properties:
        title:
          type: string
        authors:
          type: string
        abstract:
          type: string
        venue:
          type: string
        year:
          type: integer
        paperLink:
          type: ["string", "null"]
        tags:
          type: string
    ResearchPaperUpdate:
      type: object
      properties:
        title:
          type: string
        authors:
          type: string
        abstract:
          type: string
        venue:
          type: string
        year:
          type: integer
        paperLink:
          type: ["string", "null"]
        tags:
          type: string
...