# lib/api-zod/src/generated/types/researchPaperInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio API specification
 * OpenAPI spec version: 0.2.0
 */

export interface ResearchPaperInput {
  title: string;
  authors: string;
  abstract: string;
  venue: string;
  year: number;
  /** @nullable */
  paperLink?: string | null;
  tags: string;
}